requirements.txt
streamlit
PyPDF2
transformers
sentence_transformers
faiss-cpu
